requirements.txt
numpy
python-constraint
requests
networkx
pytest
pytest-cov